requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.2.0
annotated-types==0.6.0
astroid==2.12.13
async-timeout==4.0.2
asyncmy==0.2.5
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2023.11.17
charset-normalizer==2.0.8
colorama==0.4.6
dill==0.3.6
docx==0.2.4
et-xmlfile==1.1.0
fire==0.5.0
frozenlist==1.3.1
fuzzysearch==0.7.3
fuzzywuzzy==0.18.0
greenlet==2.0.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.8.0
Levenshtein==0.20.9
lxml==4.8.0
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.2
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
openpyxl==3.0.10
Pillow==9.0.1
platformdirs==2.5.4
protobuf==3.20.1
pydantic==2.5.3
pydantic_core==2.14.6
pylint==2.15.6
pyodbc==4.0.34
python-dotenv==0.21.1
pytz==2022.6
rapidfuzz==2.13.7
requests==2.26.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.31
termcolor==2.2.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.9.0
urllib3==1.26.7
wrapt==1.14.1
yarl==1.8.1
